+++ .gitignore
@@ -9,8 +9,10 @@
 *.pcap
 dpdk/x86_64-native-linuxapp-gcc/
 dpdk/build/
-example/helloworld
-example/helloworld_epoll
+example/helloworld2
+example/helloworld2_epoll
+multi-example/helloworld2
+multi-example/helloworld2_epoll
 lib/bus_if.h
 lib/cryptodev_if.c
 lib/cryptodev_if.h
