+++ config.ini
@@ -1,6 +1,6 @@
 [dpdk]
 # Hexadecimal bitmask of cores to run on.
-lcore_mask=ffffffffffffffff
+lcore_mask=1
 
 # Number of memory channels.
 channel=4
