+++ multi-example/Makefile
@@ -1,0 +1,25 @@
+TOPDIR=..
+
+ifeq ($(FF_PATH),)
+	FF_PATH=${TOPDIR}
+endif
+
+ifneq ($(shell pkg-config --exists libdpdk && echo 0),0)
+$(error "No installation of DPDK found, maybe you should export environment variable `PKG_CONFIG_PATH`")
+endif
+
+PKGCONF ?= pkg-config
+
+CFLAGS += -O -gdwarf-2 $(shell $(PKGCONF) --cflags libdpdk)
+
+LIBS+= $(shell $(PKGCONF) --static --libs libdpdk)
+LIBS+= -L${FF_PATH}/lib -Wl,--whole-archive,-lfstack,--no-whole-archive
+LIBS+= -Wl,--no-whole-archive -lrt -lm -ldl -lcrypto -pthread -lnuma
+
+TARGET="helloworld2"
+all:
+	cc ${CFLAGS} -DINET6 -o ${TARGET} main.c ${LIBS}
+
+.PHONY: clean
+clean:
+	rm -f *.o ${TARGET} ${TARGET}_epoll
